Title: 삼각형 만들기, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241212/삼각형 만들기/create-triangle.cpp
@@ -13,7 +13,7 @@ bool Parallel(int x1, int x2, int x3, int y1, int y2, int y3) {
 
 int DoubleSize() {
     bool isSatisfied = false;
-    int max_size = INT_MIN;
+    double max_size = INT_MIN;
     for(int i = 0; i < n; i++) {
         for(int j = i + 1; j < n; j++) {
             for(int k = j + 1; k < n; k++) {
@@ -21,7 +21,7 @@ int DoubleSize() {
                 int y1 = arr[i][1], y2 = arr[j][1], y3 = arr[k][1];
                 if(Parallel(x1, x2, x3, y1, y2, y3)) {
                     isSatisfied = true;
-                    int size = (x1 * y2 + x2 * y3 + x3 * y1) - (x2 * y1 + x3 * y2 + x1 * y3);
+                    double size = (x1 * y2 + x2 * y3 + x3 * y1) - (x2 * y1 + x3 * y2 + x1 * y3);
                     size /= 2;
                     max_size = max(max_size, size);
                 }
